Reorganized codebased and implemented kebab case for routing
