+++ packages/web/src/components/Example/Example.mutations.ts
@@ -1,4 +1,4 @@
-import { client } from '../../../lib/client';
+import { client } from '../../lib/api/client';
 
 export const createMessage = async (text: string) => {
   const res = await client.example.$post({
