+++ packages/web/src/components/Example/Example.queries.ts
@@ -1,4 +1,4 @@
-import { client } from '../../../lib/client';
+import { client } from '../../lib/api/client';
 
 export const getMessage = async () => {
   const res = await client.example.$get();
