+++ packages/web/src/components/ExampleNavbar/ExampleNavbar.tsx
@@ -5,9 +5,9 @@ const ExampleNavbar = () => {
   return (
     <nav className={classes.navLinks}>
       <Link to="/">Home</Link>
-      <Link to="/serverexample">ServerExample</Link>
-      <Link to="/serverexample/nestedexample">NestedExample</Link>
-      <Link to="/slugexample/slug">SlugExample</Link>
+      <Link to="/server-example">ServerExample</Link>
+      <Link to="/server-example/nested-example">NestedExample</Link>
+      <Link to="/slug-example/slug">SlugExample</Link>
       <Link to="/non-existing-slug">InvalidRoute</Link>
       <Link to="/docs">Docs</Link>
     </nav>
